+++ module/chat.js
@@ -3,7 +3,7 @@ import { DEG_Utility } from "./utility.js";
 
 export class DegenesisChat {
 
-  static async renderRollCard(rollResult, cardData) {
+  static async renderRollCard(rollResult, cardData, extraFlags = {}) {
 
     rollResult.rolls?.forEach((r) => {
       r.img = `systems/degenesis/icons/dice-faces/d${r.result}.svg`;
@@ -24,6 +24,10 @@ export class DegenesisChat {
     const html = await renderTemplate(mergedData.template, mergedData);
 
     const chatData = DEG_Utility.chatDataSetup(html, mergedData.speaker);
+
+    if (Object.keys(extraFlags).length) {
+      chatData.flags = foundry.utils.mergeObject(chatData.flags ?? {}, extraFlags);
+    }
 
     await ChatMessage.create(chatData);
   }